use bigint

+++ haf_plug_play/database/haf_sync.py
@@ -45,7 +45,7 @@ class HafSyncSetup:
         db.execute(
             f"""
                 CREATE TABLE IF NOT EXISTS public.plug_play_ops(
-                    id integer PRIMARY KEY,
+                    id bigint PRIMARY KEY,
                     block_num integer NOT NULL,
                     transaction_id char(40),
                     req_auths json,
@@ -82,17 +82,17 @@ class HafSyncSetup:
             f"""
                 CREATE TABLE IF NOT EXISTS public.plug_sync(
                     plug_name varchar(16) NOT NULL,
-                    latest_hive_rowid integer,
+                    latest_hive_rowid bigint,
                     latest_hive_head_block integer,
-                    state_hive_rowid integer
+                    state_hive_rowid bigint
                 );
             """, None
         )
         db.execute(
             f"""
                 CREATE TABLE IF NOT EXISTS public.global_props(
-                    head_hive_rowid integer,
-                    head_block_num integer,
+                    head_hive_rowid bigint,
+                    head_block_num bigint,
                     head_block_time timestamp
                 );
             """, None
@@ -108,14 +108,14 @@ class HafSyncSetup:
         # create update ops function
         db.execute(
             f"""
-                CREATE OR REPLACE FUNCTION public.update_plug_play_ops( _first_block INT, _last_block INT )
+                CREATE OR REPLACE FUNCTION public.update_plug_play_ops( _first_block BIGINT, _last_block BIGINT )
                 RETURNS void
                 LANGUAGE plpgsql
                 VOLATILE AS $function$
                     DECLARE
                         temprow RECORD;
-                        _id INTEGER;
-                        _head_hive_rowid INTEGER;
+                        _id BIGINT;
+                        _head_hive_rowid BIGINT;
                         _block_num INTEGER;
                         _block_timestamp TIMESTAMP;
                         _required_auths JSON;
